+++ src/core/client/streamers/index.ts
@@ -2,10 +2,11 @@ import * as blip from './blip';
 import * as doors from './doors';
 import * as item from './item';
 import * as marker from './marker';
+import * as markerVirtual from './markerVirtual';
 import * as object from './object';
 import * as ped from './ped';
 import * as textLabel from './textlabel';
 import * as textLabelVirtual from './textLabelVirtual';
 import * as worldNotifications from './worldNotifications';
 
-export { blip, doors, item, marker, object, ped, textLabel, textLabelVirtual, worldNotifications };
+export { blip, doors, item, marker, markerVirtual, object, ped, textLabel, textLabelVirtual, worldNotifications };